Fix warning regarding cast of std::streamsize to size_t on Windows (VS 10).
git-svn-id: https://sbia-svn.uphs.upenn.edu/projects/BASIS/trunk@1922 3c61e6fb-ae43-4039-8cf7-c895b3257837

+++ src/utilities/cxx/subprocess.cxx
@@ -643,7 +643,7 @@ bool Subprocess::communicate(std::istream& in, std::ostream& out, std::ostream& 
         while (!in.eof()) {
             in.read(buf, nbuf);
             if(in.bad()) return false;
-            write(buf, in.gcount());
+            write(buf, static_cast<size_t>(in.gcount()));
         }
 #if WINDOWS
         CloseHandle(_stdin);
